refactor: user 뷰모델, 레포지토리, 서비스 계층 역할 분리
