+++ lib/viewmodels/user/auth_view_model.dart
@@ -2,25 +2,9 @@ import 'dart:developer';
 
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:travel_muse_app/models/user/app_user_model.dart';
+import 'package:travel_muse_app/models/user/auth_state.dart';
 import 'package:travel_muse_app/repositories/user/app_user_repository.dart';
 import 'package:travel_muse_app/services/user/auth_service.dart';
-
-class AuthState {
-  AuthState({this.user, this.appUser, this.isUserNew = false});
-
-  final User? user;
-  final AppUser? appUser;
-  final bool isUserNew;
-
-  AuthState copyWith({User? user, AppUser? appUser, bool? isUserNew}) {
-    return AuthState(
-      user: user ?? this.user,
-      appUser: appUser ?? this.appUser,
-      isUserNew: isUserNew ?? this.isUserNew,
-    );
-  }
-}
 
 class AuthViewModel extends Notifier<AuthState> {
   final _authService = AuthService();
@@ -32,7 +16,7 @@ class AuthViewModel extends Notifier<AuthState> {
     return AuthState(user: currentUser);
   }
 
-  // 성공 시 상태 갱신, 실패 시 log 출력
+  // 구글 로그인
   Future<void> loginWithGoogle() async {
     final result = await _authService.signInWithGoogle();
 
@@ -50,6 +34,7 @@ class AuthViewModel extends Notifier<AuthState> {
     }
   }
 
+  // 애플 로그인
   Future<void> loginWithApple() async {
     final result = await _authService.signInWithApple();
 
@@ -57,6 +42,7 @@ class AuthViewModel extends Notifier<AuthState> {
       final user = result.data!.user;
       state = state.copyWith(user: user);
 
+      // Firestore Database에 유저 최초 등록
       await _appUserRepository.createAppUser(user!.uid);
 
       log('apple로 로그인 성공: ${result.data!.user!.uid}');
@@ -85,12 +71,9 @@ class AuthViewModel extends Notifier<AuthState> {
     }
   }
 
+  // 로그아웃
   Future<void> logout() async {
     await _authService.signOut();
     state = AuthState(); // 초기화
   }
 }
-
-final authViewModelProvider = NotifierProvider<AuthViewModel, AuthState>(
-  () => AuthViewModel(),
-);
