+++ lib/views/home/widgets/info_banner.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:intl/intl.dart';
 import 'package:travel_muse_app/providers/plan/calendar_location_provider.dart';
-import 'package:travel_muse_app/viewmodels/user/auth_view_model.dart';
+import 'package:travel_muse_app/providers/user/auth_view_model_provider.dart';
 import 'package:travel_muse_app/views/plan/plan/schedule/schedule_page.dart';
 
 class InfoBanner extends ConsumerWidget {
