+++ lib/views/user/onboarding/onboarding_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:travel_muse_app/constants/app_text_styles.dart';
-import 'package:travel_muse_app/viewmodels/user/profile_view_model.dart';
+import 'package:travel_muse_app/providers/user/profile_view_model_provider.dart';
 import 'package:travel_muse_app/views/user/onboarding/terms_agreement_bottom_sheet.dart';
 import 'package:travel_muse_app/views/user/onboarding/widgets/edit_birth_date.dart';
 import 'package:travel_muse_app/views/user/onboarding/widgets/select_gender.dart';
