+++ lib/views/user/onboarding/widgets/check_duplicate_button.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:travel_muse_app/viewmodels/user/profile_view_model.dart';
+import 'package:travel_muse_app/providers/user/profile_view_model_provider.dart';
 import 'package:travel_muse_app/views/user/onboarding/widgets/duplicate_button_themes.dart';
 
 class DuplicateCheckButton extends ConsumerWidget {
