+++ lib/views/widgets/edit_profile_image.dart
@@ -5,7 +5,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:travel_muse_app/constants/app_colors.dart';
 import 'package:travel_muse_app/constants/app_text_styles.dart';
-import 'package:travel_muse_app/viewmodels/user/profile_view_model.dart';
+import 'package:travel_muse_app/providers/user/profile_view_model_provider.dart';
 
 class EditProfileImage extends ConsumerStatefulWidget {
   const EditProfileImage({super.key, required this.size});
